rename: 패키지 명칭변경 (풀네임으로 변경)
